+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-- Added support for Legacy4J 1.9.
-- Legacy4J: Added support for the new built in Change Skin screen.
+- The Paper Doll and Change Skin button can now be freely repositioned by holding `Shift + Right-Click`. As a result, the `Paper Doll Position` option has been removed.
+- Improved Paper Doll rendering.
 
 **Full Changelog**: https://github.com/BrandonItaly/Bedrock-Skins/commits/master